move fn to src/event-helper.js
